pyproject.toml
[project]
name = "learning-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.1.1",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.8",
    "semantic-chunkers>=0.0.10",
    "semantic-router>=0.0.68",
]
